Remove emacs/xemacs handling from configure.ac
This makes configure prefer emacs over xemacs, it used to be the other
way around.

Verify that something that looks like Emacs is found by
AM_PATH_LISPDIR and notify the user with an error otherwise.

AC_SUBST(EMACS) is not needed, already done by AC_CHECK_PROG.
The --with-emacs and --with-xemacs options are not mutually exclusive,
 and "configure EMACS=xemacs" works better anyway.
AM_PATH_LISPDIR already does the equivalent of the AC_CHECK_PROG
 statement (except it prefers emacs over xemacs), so that can go too.
AM_PATH_LISPDIR also handles EMACS=t, so that test can also be removed.