Don't shadow case-fold-search with mmm-ify argument
The mmm class property :case-fold-search is passed as a keyword
argument to mmm-ify.  But that argument would shadow the dynamic
variable with the same name.  Use the name case-fold-search-arg with a
:case-fold-search keyword to avoid that problem.

The argument was used instead of a let-bound variable, so add a let
binding of case-fold-search around the function body.